requirements.txt
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
dataclass-wizard==0.23.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
paho-mqtt==2.1.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pytest==8.3.3
pytest-cov==5.0.0
requests==2.32.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
